requirements.txt
# PlantIQ AI Brain - Dependencies
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# ML & Data Science
scikit-learn==1.5.2
xgboost==2.1.1
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1
joblib==1.4.2

# Time-series (lightweight alternative to Prophet)
statsmodels==0.14.2

# API & Middleware
python-multipart==0.0.9
httpx==0.27.0

# Utilities
python-dateutil==2.9.0
